Implement User Registration and Authentication

+++ Day 87 - Web Development - Todo List Webapp/app.py
@@ -1,23 +1,35 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for, flash
 from flask_sqlalchemy import SQLAlchemy
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, HiddenField, BooleanField
-from wtforms.validators import DataRequired
+from wtforms import StringField, SubmitField, HiddenField, BooleanField, PasswordField
+from wtforms.validators import DataRequired, Length
 from flask_bootstrap import Bootstrap
+from flask_login import LoginManager, UserMixin, login_user, login_required, current_user, logout_user
+from werkzeug.security import generate_password_hash, check_password_hash
 
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///tasks.db'
 app.config['SECRET_KEY'] = 'c7e3bbf68f5c4a8004fc08ff8858117f'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 
 bootstrap = Bootstrap(app)
+login_manager = LoginManager(app)
+login_manager.login_view = 'login'
 
+
+class User(UserMixin, db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(20), unique=True, nullable=False)
+    password = db.Column(db.String(80), nullable=False)
+    tasks = db.relationship('Task', backref='author', lazy=True)
 
 class Task(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     content = db.Column(db.String(200), nullable=False)
     done = db.Column(db.Boolean, default=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
 
 class TaskForm(FlaskForm):
     task = StringField('Task', validators=[DataRequired()])
@@ -25,30 +37,85 @@ class TaskForm(FlaskForm):
     done = BooleanField('Done')
     submit = SubmitField('Update Task')
 
+class RegistrationForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired(), Length(min=4, max=16)])
+    password = PasswordField('Password', validators=[DataRequired()])
+    submit = SubmitField('Sign Up')
+
+class LoginForm(FlaskForm):
+    username = StringField('Username', validators=[DataRequired(), Length(min=4, max=16)])
+    password = PasswordField('Password', validators=[DataRequired()])
+    submit = SubmitField('Login')
+
+
 #----- run once to create db -----#
 # with app.app_context():
 #     db.create_all()
 
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(int(user_id))
+
+
 @app.route('/')
 def index():
-    tasks = Task.query.all()
-    form = TaskForm()
-    return render_template('index.html', tasks=tasks, form=form)
+    if current_user.is_authenticated:
+        tasks = Task.query.filter_by(user_id=current_user.id)
+        form = TaskForm()
+        return render_template('index.html', tasks=tasks, form=form)
+    else:
+        return render_template('home.html')
+    
+@app.route('/register', methods=['GET', 'POST'])
+def register():
+    form = RegistrationForm()
+    if form.validate_on_submit():
+        hashed_password = generate_password_hash(form.password.data, method='pbkdf2:sha256', salt_length=8)
+        new_user = User(username=form.username.data, password=hashed_password)
+        db.session.add(new_user)
+        db.session.commit()
+        flash('Your account has been successfully created! You can now log in.', 'success')
+        return redirect(url_for('login'))
+    return render_template('register.html', form=form)
+
+@app.route('/login', methods=['GET', 'POST'])
+def login():
+    form = LoginForm()
+    if form.validate_on_submit():
+        user = User.query.filter_by(username=form.username.data).first()
+        if user and check_password_hash(user.password, form.password.data):
+            login_user(user)
+            return redirect(url_for('index'))
+        else:
+            flash('Incorrect Username or Password', 'danger')
+    return render_template('login.html', form=form)
+
+@app.route('/logout')
+@login_required
+def logout():
+    logout_user()
+    return redirect(url_for('index'))
+
 
 @app.route('/add', methods=['POST'])
+@login_required
 def add_task():
     form = TaskForm(request.form)
-    if form.validate_on_submit:
+    if form.validate_on_submit():
         new_task_content = form.task.data
         new_task_done = form.done.data
-        new_task = Task(content=new_task_content, done=new_task_done)
+        new_task = Task(content=new_task_content, done=new_task_done, user_id=current_user.id)
         db.session.add(new_task)
         db.session.commit()
     return redirect(url_for('index'))
 
 @app.route('/edit/<int:task_id>', methods=['GET', 'POST'])
+@login_required
 def edit_task(task_id):
     task = Task.query.get_or_404(task_id)
+    if task.author != current_user:
+        abort(403)
+
     form = TaskForm(request.form, obj=task)
 
     if request.method == 'POST' and form.validate_on_submit():
@@ -60,10 +127,15 @@ def edit_task(task_id):
     return render_template('edit_task.html', form=form, task=task)
 
 @app.route('/delete/<int:task_id>')
+@login_required
 def delete_task(task_id):
     task = Task.query.get_or_404(task_id)
+    if task.author != current_user:
+        abort(403)
+
     db.session.delete(task)
     db.session.commit()
+    flash('Task deleted successfully!', 'success')
     return redirect(url_for('index'))
 
 if __name__ == '__main__':
